code is fully rewritten, refactoring made, module structure made for easy integrating new cdes

+++ src/main/java/ru/mauveferret/GUI.java
@@ -36,7 +36,7 @@ public class GUI extends Application {
         System.out.println("?!");
     }
 
-    public void showGraph(int spectra[], int E0, double dE, String name)
+    public void showGraph(int spectra[], double E0, double dE, String name)
     {
         Stage dialogStage = new Stage();
         dialogStage.setTitle("SCATTER ONE LOVE");
